Update patch set 3
Patch Set 3: Code-Review+2

I could also see users wanting it to work as an override instead of an addition, to select an alternate config which is not the one shipped in the default location. I assume there's a good reason you didn't choose to go that route with it, and I'm just not thinking of it.

Patch-set: 3
Reviewer: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2