Update patch set 3
Patch Set 3:

I started implementing it that way, but there's one behavior change which made it difficult to continue down that path: we stop after reading the first of the 4 standard config locations.  That means that if we implemented it as just an override of the usual path precedence, we would not be able to have *additional* locations.  (E.g., if we had a hypothetical "config-paths: [zuul.yaml, extra.yaml]" option, then we would end up only loading from one of those, whereas the goal is to load from both.

I see 3 ways forward:

1) This change -- the minimal additional functionality, not interfering with the original behavior.
2) Declare the original behavior a mis-feature, and change it so that we do not stop on the first found.  That would allow us to implement this feature as an override of the path precedence.  Hopefully most people don't have zuul.yaml and .zuul.yaml, and we can get away with just doing this as a minor point release with an upgrade note.
3) Stick with this change, and implement yet another config option in the future to allow overriding the standard first-found locations.

Patch-set: 3