Update patch set 3
Patch Set 3:

Yes, let's call that #4.  The question is whether first-found with alternate paths is useful.  I think we would have that with #3, but with #4 we would only have the "load everything found" behavior we get with the extra-config-paths option in this change.

(I guess we could also probably add a #5 which is this change but implement the extra-config-paths as a first-found list, but one that just operates sequentially after the built-in first-found list.  That would still meet the use case I have for this, which is to load stuff from files in zuul-test.d/ in the zuul/zuul-jobs repo.)

Patch-set: 3
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0