Update patch set 3
Patch Set 3:

As it stands, this change satisfies the need we have for testing the zuul-jobs repo in OpenDev's Zuul, so I'm in favor. It would be good to hear if there are others for whom the alternate described behaviors solve different use cases they have (it sounded like at least Clint was interested in something along these lines but not sure if this takes care of it already), and I guess more importantly whether this feature makes it harder/more confusing to address those cleanly.

Patch-set: 3