Update patch set 7
Patch Set 7:

Based on the CephCluster CRD design, I'm actually now leaning towards thinking we should make connection info be part of the Zuul CRD and not it's own. It seems like we should have a Zuul and a Nodepool, since those are the resources we want the k8s scheduler to manage. A ZuulConnection doesn't feel right as a top-level resource, since what it really is is a facet of the config of a Zuul. I'll push up a revision with that change made.

Patch-set: 7
Label: Verified=0