Build layout of most recent non-live item
The recent change to diff job configs in order to run modified jobs
even if file matchers don't match compares the job config in the
change under test to the most recent layout ahead in the queue, or
the running tenant layout if none, in order to determine if a job's
config has changed.  In an independent pipeline with multiple
dependent changes, we did not generate layouts for non-live items,
therefore the only layout available to diff was the running tenant
layout.  This would cause changes to run too many jobs in that
situation (while they would run the correct jobs in a dependent
pipeline since layouts of items ahead were available).

To correct this, always generate the layout for the last non-live
item in a series, so that the first live item can diff against it.

Change-Id: Id533772f35ebbc76910398e0e0fa50a3abfceb52