Use urllib.parse for manipulating client urls
The current code appends a port if one is defined, but if self.url
had a trailing path component, this would put the port in the wrong
place. Instead, parse the url into bits, update the port, and
recompose a new URL.

It would be nice if there was a nicer way to do this, but such is life.

Change-Id: I282cd69662ea2fd533745a0ca6de3d8ddd685842