Handle existing broken config in job updates
If Zuul was unable to freeze the job graph of the previous layout,
then it would throw an exception when calculating whether jobs have
changed.  This is a problem if it does so on a change which fixes
an error in the current tenant config, as it means the fix can not
merge.

To handle this case, we need to catch that exception, and then provide
a default value indicating whether the job config has been updated
or not.  Because returning True could cause every job with a file
matcher to run unecessarily, the safer default is False (so the job
relies on the file matcher alone).

Change-Id: I8a72e57e068b073e37274cdabeacffc9daee2e94