Update patch set 4
Patch Set 4:

(1 comment)

Could we also add mention of making this configurable? I think allowing circular dependencies should be off by default. I'm not sure if this is a pipeline option or a tenant option - I'd guess tenant. I know, for instance, that the openstack tenant of opendev would want to keep the setting disabled... but I'm guessing in multi-tenant zuuls it's possible one tenant would want it off and one would want it on, so it seems it shouldn't be a global setting.

Patch-set: 4
Label: Verified=0