Update patch set 4
Patch Set 4: Code-Review-1

I'd like to reserve fixed tuning options like these as a last resort so that we don't overwhelm operators with tuning variables.  Is there a way we can provide a sensible default or heuristic here?

The 2x seems to work for some cases -- where do you think it starts to break down (ie, with how many cpus does it become too much)?  Maybe we can use 2x up to some point and then 1x after that point.  (eg, if we set the threshold at 8, then an 8vcpu machine is 8*2.5*2, but a 12 vcpu machine is 8*2.5*2+4*2.5.  That's just an idea off the top of my head.)

Patch-set: 4
Label: Code-Review=-1