Update patch set 4
Patch Set 4:

I agree that the default threshold/heuristic should fit as many installations as possible if not all. We could improve this probably by limiting the number of cpus that are taken into account to maybe 4. This would emphasise that job startup is mostly IO limited.

However I cannot think of a generic way of determining a threshold that fits every system (specificly ours). The reason is that the job startup times are highly dependent on various system characteristics like IO performance, typical type of jobs, number of executors. Especially large jobs with multiple large repos (which are probably uncommon in many deployments) quickly become a problem in our environment which result in >20 minutes job startup times if we start too many jobs in parallel. In our case the behavior is completely completely independent of the number of CPUs but the system is highly dependent of the specific IO performance.

So in our specific system with our typical workload things start to break down with a limit of more than 4 or 5 jobs (despite having 16 cpus). So I think that specifically our system really needs methods to fine tune this.

So I'd suggest as a compromise we could tune the current algorithm (maybe as suggested above) to fit more/most systems because currently for more cores this limit becomes actually not really a limit. And further make this overridable so we can tune it down so it doesn't break down for our workload.

Patch-set: 4
Label: Verified=0