Update patch set 5
Patch Set 5:

As an FYI, another requirement not mentioned in this spec is the need to change how autohold requests are handled. These requests are currently held in memory. Scaling out the scheduler would mean we'd need to alter that to a shared storage. The following changes, once merged, will handle that for us (just mentioning for completeness):

https://review.opendev.org/#/q/project:zuul/zuul+branch:master+topic:autohold-revamp

Patch-set: 5
Label: Verified=0