Allow projects to borrow ssh keys
If a final job is run in a post-review pipeline for one of its
allowed-projects, then let the allowed-project borrow the ssh
key of the project where the job is defined.

The idea is that if the defining job says the other project can
run a deployment job, it is also saying it can use its credentials
to do so.

The use case that drove this is a case where the OpenDev admins
want to define a job to update the nameservers in system-config,
but also run it on changes to the zonefiles, which are in dedicated
repos which may have other collaborators working on them.  We don't
want to add the project key of the zonefile repos to our bastion host,
but we do want to allow the one job which updates the nameservers to
use the system-config key.

Change-Id: I55a67c5a848874fbd826c80d55517b4304b0a9b7