Update heuristing of parallel starting builds.
An executor is accepting up to twice as many starting builds as defined
by the load_multiplier option. This is limited to 4 CPU/vCPU count.
After that the executor is accepting only up to as many starting builds
as defined by the load_multiplier (also up to half as many).

Change-Id: I8cf395c41191647605ec47d1f5681dc46675546d