WIP: Add manifest tree view to build page
This doesn't use the right metadata yet (waiting for the generate
manifest role to land).  It also doesn't show or link to any file
contents yet.

But it does fetch the manifest and show the tree view.

Try it with build 129b4b885e9d449c9edc173008e67c72 in the opendev
zuul tenant.

Change-Id: Ib04d013b4118005ba66a91d2bec0b0c429d12863