WIP Use cherrypy_cors to set cors headers
Instead of having to set them in every method, set them globally
using the cherrypy_cors module.

Also, add the tool to after_error_response so that we set the
headers on the error responses too.

This is WIP because we should probably wrap our heads around
how this interacts with the auth stack. The cors tools can be used
as decorators as well.

Change-Id: Iddd730c132bbc567a741dc45ceb4dac13f481af8