Improve SQL query performance in some cases
The query for builds-joined-with-buildsets is currently optimized
for the case where little additional filtering is performed.  E.g.,
the case where a user browses to the builds tab and does not enter
any search terms.  In that case, mysql needs a hint supplied in
order to choose the right index.

When search terms are entered which can, due to the presense of
other indexes, greatly reduce the working set, it's better to let
the query planner off the leash and it will make the right choices.

This change stops adding the hint in the cases where a user supplies
a search term that matches one of the indexes on the build or
buildset table (notable exception: job_name because it is difficult
to generalize about that one).

It also adds an additional index for build and buildset uuids,
which should provide excellent performance when searching for
only those terms.

Change-Id: I0277be8cc4ba7555c5e6a9a7eb3eed988a24469c