Update patch set 5
Patch Set 5:

> Where is the redux store stored?  On disk?  In ram?  I'm still
 > pretty uncomfortable putting lots of large files in it.  I think if
 > folks are going to switch back and forth, they're more likely to
 > open up multiple tabs anyway?  (caching only helps us if they view
 > file A, then B, then return to A)

The redux store resides in ram and it is unique per tab by default (there are optional middleware to persist it on disk).

I agree that large files shouldn't be cached and we should force opening them in another view/tab. But for small file, if we cache them then we can display them in the tree view, which may help significantly the browsing.

Patch-set: 5