Update patch set 5
Patch Set 5:

Okay, that makes sense.  I'm skeptical that browsing any of the log files in the tree view is going to be a good experience, but I think we can give it a try and would love to be surprised by a good new UI.

I propose that we merge this series as-is, since right now it matches the spec and doesn't store more than one file (meaning that we're using the minimum amount of ram we can use at the moment).  If any of us is interested in exploring further changes like you describe, it's easy enough to improve the caching so that it's per-file (and add some kind of size filter or expiration mechanism).  We can evaluate those on their own.  Even if we might be able to come up with something better in the future, I think this series is pretty good on its own.

Patch-set: 5