Always record the log_url in sql reporter
The build page needs the actual log_url returned by the job (without
any modification from success_url or failure_url) in order to create
links to the log site.

The reported success/failure URL isn't as important in this context,
and I suspect their days are numbered once we require the SQL
reporter and report the link to the build page instead.  So we just
won't record those in the DB.  If we feel that they are important,
we can add new columns for them.

Also, ensure it has a trailing / so that API users (including the JS
pages) are not annoyed by inconsistent data.

Change-Id: I5ea98158d204ae17280c4bf5921e2edf4483cf0a