Stack listview in build console
This stacks the "Pre playbook" text over the playbook path, which
may be considered a visual improvement.

Note that we can't mix-and match styles, so if we stack this,
then if we start using both header and description text in other
rows, they will also be stacked.  Currently, we only use the
header text in the other rows, so it's not presently an issue.

Change-Id: Iee0a2970ec46ca7d7ea3293ae43d0570e73cd3a7