Flatten host tasks into tasks on build page
This removes a level of nesting of "hosts" underneath "tasks" on
the build page console.  Tasks run on multiple nodes are simply
repeated (with a different node name in the header).  This removes
a level of potentially unintutive nesting, and simplifies the output
on single-node jobs, but may increase visual clutter on multinode
jobs due to the repeated tasks.

Change-Id: I532df1ff5ad07e034dcbe8dcbc88166d764f7915