Update patch set 2
Patch Set 2:

Even if the play has complex host list, what will render here is the individual hosts that match.  So if we had a play with "hosts: z*" we would see:

* Install something (zuul.opendev.org)
* Install something (ze01.opendev.org)

You might be able to find an example of a multinode job and see it in action.

Patch-set: 2