Don't always show expansion option on build console
If a task has no interesting output, don't show the expansion
caret.  In order to make this look right, we add in a blank
leftContent section in place of the caret, and tweak the css
for that so it takes the same amount of space.

Since the row isn't clickable in that case, make sure that we
use the pointer icon over the details button, and while we're
at it, make sure we use the default cursor over the additional
info items which are not clickable.

Change-Id: I5d5691ef480af932fab0d592eacfd260426352dd