Add permalinks to task detail popup
This adds a link icon to the title of the modal dialog that opens
when the task status is clicked.  It uses the URL hash to supply
a path to the task which will cause all the parent tasks to be
expanded and the dialog open when a user visits the url with the
hash.

Change-Id: I3cf5b7115b775de103b9af7986ac0b8318688b42