Consistent handling of ansible venv install dirs
The manage_ansible.py script/command only installs ansible venvs to
sys.exec_prefix/lib/zuul/ansible which is likely /usr/lib/zuul/ansible
on most systems. The executor server overrides the installation
directory to be /var/lib/zuul/ansible-bin, but this override only
applies if uses haven't previously run the manage_ansible.py script.

This happens because the internal ManagedAnsible class checks all paths
it knows about for the venvs and uses the first it finds. In those cases
it finds the /usr/lib/zuul/ansible venvs even though it was told to use
/var/lib/zuul/ansible-bin and it uses the /usr paths as they exist.

We clean this up in two ways. First we stop providing an overridden
default value in the executor server. It should just use the
ManagedAnsible defaults so that it aligns with manage_ansible.py better.
Second we update ManagedAnsible venv search dir ordering to prefer
overridden paths so that they will be used over the defaults if
selected.

Change-Id: If84a92bc542e24eedac09f3dc9fd2954db43734c