Update patch set 1
Patch Set 1:

I'm curious about this... ultimately this is just code organization, right?  These functions are still called within the rendering call stack, so it doesn't really change whether the Console is suitable to be a pure component.  Given that they are only used in the Console file, can't they continue to stay there?

Also, my understanding is that these are unrelated to whether a component is 'pure' or not.  It only needs to render the same output given the same inputs.  That doesn't mean they can't call functions to do that.

Patch-set: 1