Update patch set 1
Patch Set 1:

I read your message after I wrote mine, so let me follow up with another direct question -- why shouldn't we have functions which are only used by a container in the same file as the container?   After all, there are similar functions which could be methods on the class, and they would necessarily be in this file.  These don't need to be methods, but they're still only useful in this context.

Patch-set: 1