Implement push job in merger
This is the actual implementation of the direct-push as a new merger
job. This job can the be used by different components (e.g. the
GerritReporter) to push the changes to the remote when everything was
successful.
To ensure that we push the same state we used for the validation, the
push job reuses the merge methods to create the same state of the repo.

NOTE: The last part was an assumption from my side, but following
Clark's comment that might not be the case when using the _mergeItem()
method directly.

Change-Id: I48431163892e9e1e438b910961218aa2dfd79b49