JS: Break log viewer out of the panel
Confining the log view to a patternfly panel means that the inner
panel needs to be horizontally scrolled, but to get to the horizontal
scrollbar, you have to go to the bottom of the log.  Instead, render
the logs outside of any panels, with no overflow handling and let
the browser do what it does natually.

Change-Id: I8258ecf6feb036ee15aae3e314363376d71292de