WIP: super hacky demo of logfile under the manifest
This is basically a demo of what it would look like if we put the
log file rendering on the same page as the manifest.  This makes
it easier to switch between different files quickly.

Lots of stuff doesn't work right yet, this is just for initial
evaluation.  You should be able to navigate to the logs tab of
a build, and click one of the black (not blue) filenames to see
this in action.

Change-Id: I81de107dfe32c48e85fe6a15d19c253e2045aa37