Update patch set 2
Patch Set 2:

actually, this works when the page load, but when lines are clicked then the browser moves the anchor at the top, which seems like the default behavior. it seems like one solution would be to not use anchor. We need that anyway to be able to select multiple lines.

Patch-set: 2
Reviewer: Gerrit User 9311 <9311@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0