Stop executor if log streamer dies
We've run into cases where the log streamer might be killed (because
OOMKiller). When this happens the executor ceases to fully function:
users can't get console logs for jobs run on this executor. To avoid
this have the entire executor stop running then system level service
management can decide if the service should be started again.

Change-Id: Ie36f586fa895902a222ab61a5f038d6e670a0f49