Restart log streamer if it dies
We've run into cases where the log streamer might be killed (because
OOMKiller). It would be nice if zuul continued to function after this so
have zuul restart the log streamer process if it dies.

To do this we add a new log streamer manager thread that can waitpid
periodically to see if its child has died and restart it if so.

Change-Id: I0788053abcc64f082dcd63af4dbe7a23e78706ad