Update patch set 1
Patch Set 1:

I've debugged this issue before; I agree it certainly can be confusing when you're not looking for it!

We're peaking in memory usage and it might be fine to handle the OOM kill as a one-off ... but then again we might be just restarting the thing again to it's certain death.  So perhaps a bit of hysteresis might be appropriate?

Again, just as someone who jumped in to debug this, perhaps it would be better to monitor the thread and kill the whole executor with a sensible message if it disappears?  Clearly something the executor as a whole was doing at the time got us into the situation, so if the whole thing stops  there might be some chance you can try and isolate it a bit.  something monitoring zuul-executor as a whole (systemd, etc) could handle the wholesale restart if required?

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0