Update patch set 1
Abandoned

I've become convinced that we should not encourage folks to use the Python logging module configuration. In most cases its only use is to turn on DEBUG level logging, and for that purpose it's a very complex and error-prone solution. Instead let's de-emphasize log config, and add a feature like https://review.opendev.org/635649 to toggle it from the command-line.

Patch-set: 1
Status: abandoned