This is a stub for a bug to help us find and fix it
What appears to happen is zuul decideds a repo does not need to be
updated because the commit or its parent (something) already exists in
the repo. If it is the case that a new branch has been created from a
shared parent then we don't update to get the new branch and when that
happens we don't select that branch on the remote test node. The commit
is there we just don't check it out so it doesn't get tested.

This adds a stub which totally doesn't work to add change related ref
checking in addition to change related commit checking. But its late so
I'm calling it a day.

Change-Id: If58168a45722f53676c764c2cc5549434e3a6d6a