[wip] Test and expand documentation for executor-only jobs
When running an executor-only job (i.e. with an empty nodeset) the
inventory is empty and thus the playbook runs with the "implicit
localhost".

There are a number of caveats to watch out for when running like this;
for example hostvars.values() will not work hostvars['localhost'] will
(see [1]).

This updates the current discussion about executor-only jobs with some
more information and a pointer to the Ansible page on implicit
localhost.

For testing, we add a executor-only job and ensure that the inventory
is left blank.

I5e23f330476f064acf3cb87f746c5d3193cce274

[1] https://github.com/ansible/ansible/issues/61513

Change-Id: I8fd2f932290e49da5a3605737e8940425cd092f4