executor: always setup an inventory, even for blank node list
I found this when watching the log streaming for one of the new
promote pipeline jobs in OpenDev.  I noticed sending the logs to
logstash was failing (this happens *after* the log upload, so you
don't see it unless you're reading the server logs or watching the
stream).  After diagnosing with
I20d1d99a48b7a173ab6c792dc3508c27b8047f6a it became clear that
"hostvars" was not set, causing the role to fail [1]:

   self.zuul = list(host_vars.values())[0]['zuul']
 IndexError: list index out of range

"host_vars" is passed in from "hostvars", and this could only mean it
was undefined ... a weird thing for an Ansible builtin variable.

After digging in the logs I found the log line that helped me crack
the case ...

 Ansible command: blah blah
 Ansible output: b'... ansible/inventory.yaml did not meet host_list requirements, check plugin documentation if this is unexpected'
 Ansible output: b'... ansible/inventory.yaml did not meet script requirements, check plugin documentation if this is unexpected'
 Ansible output: b' [WARNING]: provided hosts list is empty, only localhost is available. Note'
 Ansible output: b"that the implicit localhost does not match 'all'"

The first two are red-herrings; confusing messages that mean "the
host_list and script inventory plugins couldn't understand this yaml
inventory file" but do not actually matter (I think Ansible 2.8 has
fixed this).

But the "hosts list is empty" is a problem.  I broke into the job
while it was running to see the inventory, and sure enough it was
blank.

 root@ze01:/var/lib/zuul/builds/...# cat ./ansible/inventory.yaml
 all:
   hosts: {}
   vars:
  ...

Sure enough, if you run a playbook like this "hostvars" will not be
defined by Ansible [2].

Then I realised this did *not* happen on the gate job, only the
promote pipeline; the difference there is that the job is "executor
only", i.e. the nodes list is empty [3].

So following a test-driven development model, I dug around in the
inventory test to make a new test that didn't define any nodes.  I ran
it, expecting to hit a new assert that we had no hosts in the
inventory ... but it didn't.

Then I realised that in the RecordingAnsibleJob we use in the test
cases, we override getHostList() and, if the host list seems empty, we
add the localhost host with a "local" connection.

So I believe the solution to this is to move this logic up into the
executor itself.  If it sees a blank node list; i.e.

 - nodelist:
    nodes: []

it should automatically configure just the localhost connection.

This does that, and adds a test-case that an executor-only job gets
the localhost host defined.

[1] https://opendev.org/opendev/base-jobs/src/commit/597135d5281b0de6bf9551b8dadef90014bb19a6/roles/submit-log-processor-jobs/library/submit_log_processor_jobs.py#L82
[2] Maybe it's an Ansible bug that it is using the implicit localhost
    but not filling out "hostvars", I don't know yet.  See:
      https://github.com/ansible/ansible/issues/61513
[2] https://opendev.org/opendev/base-jobs/src/commit/597135d5281b0de6bf9551b8dadef90014bb19a6/zuul.d/jobs.yaml#L357

Change-Id: I8fd2f932290e49da5a3605737e8940425cd092f4