Update patch set 2
Patch Set 2: Code-Review-1

We don't add localhost for a reason. The problem with this is that many playbooks out there run on 'all' hosts and expect to run only on remote hosts and not localhost which is the executor and restricted in untrusted playbooks. This relies on the implicit defined localhost which would get broken by this change.

In case you're interested in the zuul vars you don't need to take them explicitly from a specific hostvar as they're injected as host independent vars.

Patch-set: 2
Label: Code-Review=-1