Update patch set 2
Patch Set 2:

> The problem with this is that many playbooks out there run on 'all' hosts and expect to run only on remote hosts and not localhost which is the executor and restricted in untrusted playbooks. 

So now I re-read the change title, I release it is misleading.  It should read more like "setup a localhost inventory for a blank node list" ... i.e. this only adds the inventory for the case

 nodeset:
   nodes: []

which is otherwise not created.  It doesn't *always* add localhost which I agree would be incorrect.

In that case, it seems you can only be expecting playbooks that will run on localhost and be trusted enough to do what they're doing?

I am still unsure if this is a work-around for a bug (https://github.com/ansible/ansible/issues/61513) or expected behaviour of ansible, however.

Patch-set: 2