Update patch set 2
Patch Set 2:

> > The problem with this is that many playbooks out there run on
 > 'all' hosts and expect to run only on remote hosts and not
 > localhost which is the executor and restricted in untrusted
 > playbooks.
 > 
 > So now I re-read the change title, I release it is misleading.  It
 > should read more like "setup a localhost inventory for a blank node
 > list" ... i.e. this only adds the inventory for the case
 > 
 > nodeset:
 > nodes: []
 > 
 > which is otherwise not created.  It doesn't *always* add localhost
 > which I agree would be incorrect.
 > 
 > In that case, it seems you can only be expecting playbooks that
 > will run on localhost and be trusted enough to do what they're
 > doing?
 > 
 > I am still unsure if this is a work-around for a bug
 > (https://github.com/ansible/ansible/issues/61513) or expected
 > behaviour of ansible, however.

For the executor only jobs case I think this is ok. The reason for that is even in the case of an untrusted executor only job it could add host localhost and we rely on other protections against that bein exploited.

It would be good for others to confirm my understanding though.

Patch-set: 2
Reviewer: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0