Update patch set 32
Patch Set 32:

@Fabien Boucher: I was not able to see CODEOWNERS relevant output from the https://developer.github.com/v3/pulls/review_requests/#list-review-requests endpoint. From what I see in my tests it is an API to the "Reviewers" box you see on the top right corner in Github Web interface on a PullRequest - requested reviewers. Some of those will be suggested by Github based on the CODEOWNERS file but it does not mean that all are actually needed. Also they may be, and often are, teams. 

For example a CODEOWNER file:

  * @whole-team
  /some/path @some-specialists

will suggest in the above API both the "whole-team" and "some-specialists". In case there is a change in, e.g. /some/path/test.txt, it is enough that someone from the team "some-specialists" reviews it. The above API does not seem to check that. Even after that team gives a positive review the endpoint returns "whole-team". This would be then interpreted as if a review is still needed by Zuul, although Github is already satisfied.

Or did I miss something?

Patch-set: 32
Label: Verified=0